rename convert2bedRMod into bedRMod
